# docs/swagger.yaml
basePath: /
definitions:
...
  position.Position:
    properties:
      created:
        type: string
      created_by:
        type: string
      id:
        type: integer
      name:
        type: string
      organization_id:
        type: integer
      status:
        type: integer
      updated:
        type: string
      updated_by:
        type: string
    type: object
  position.PositionNew:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - name
    - status
    type: object
  project.Project:
    properties:
      created:
        type: string
      created_by:
        type: string
      id:
        type: integer
      name:
        type: string
      organization_id:
        type: integer
      status:
        type: integer
      updated:
        type: string
      updated_by:
        type: string
    type: object
  project.ProjectNew:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - name
    - status
    type: object
  response.ErrorRes:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.ListRes:
    properties:
      count:
        type: integer
      data: {}
      page_id:
        type: integer
      page_size:
        type: integer
    type: object
  response.SuccessRes:
    properties:
      data: {}
    type: object
host: 0.0.0.0:8080
info:
  contact:
    email: lxx0103@yahoo.com
    name: Lewis
  description: API for zoho-center.
  title: zoho-center API
  version: "1.0"
paths:
...
  /positions:
    get:
      consumes:
      - application/json
      operationId: "28"
      parameters:
      - description: 页码
        in: query
        name: page_id
        required: true
        type: integer
      - description: 每页行数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 职位编码
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ListRes'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/position.Position'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 职位列表
      tags:
      - 职位管理
    post:
      consumes:
      - application/json
      operationId: "29"
      parameters:
      - description: 职位信息
        in: body
        name: position_info
        required: true
        schema:
          $ref: '#/definitions/position.PositionNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/position.Position'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 新建职位
      tags:
      - 职位管理
  /positions/:id:
    get:
      consumes:
      - application/json
      operationId: "30"
      parameters:
      - description: 职位ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/position.Position'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 根据ID获取职位
      tags:
      - 职位管理
    put:
      consumes:
      - application/json
      operationId: "31"
      parameters:
      - description: 职位ID
        in: path
        name: id
        required: true
        type: integer
      - description: 职位信息
        in: body
        name: position_info
        required: true
        schema:
          $ref: '#/definitions/position.PositionNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/position.Position'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 根据ID更新职位
      tags:
      - 职位管理
  /projects:
    get:
      consumes:
      - application/json
      operationId: "5"
      parameters:
      - description: 页码
        in: query
        name: page_id
        required: true
        type: integer
      - description: 每页行数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 项目编码
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ListRes'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project.Project'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 项目列表
      tags:
      - 项目管理
    post:
      consumes:
      - application/json
      operationId: "6"
      parameters:
      - description: 项目信息
        in: body
        name: project_info
        required: true
        schema:
          $ref: '#/definitions/project.ProjectNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/project.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 新建项目
      tags:
      - 项目管理
  /projects/:id:
    get:
      consumes:
      - application/json
      operationId: "7"
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/project.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 根据ID获取项目
      tags:
      - 项目管理
    put:
      consumes:
      - application/json
      operationId: "8"
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 项目信息
        in: body
        name: project_info
        required: true
        schema:
          $ref: '#/definitions/project.ProjectNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/project.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 根据ID更新项目
      tags:
      - 项目管理
...
swagger: "2.0"